refactor: extract delete-confirm overlay (#357)
Fifth slice of the ui.rs decomposition. Smallest overlay so far
(~22 lines): the y/l/n delete confirmation popup that adapts its
prompt for outgoing vs incoming messages.

ui/mod.rs: 4800 -> 4778 lines.
